interfaces: store Connections in each Slot and Plug
This patch adds a way to store connections inside each Slot and Plug.
While the repository doesn't actually use it in any way it is a
convenient data format to have so that we don't need to create similar
structures in the deamon code. The added advantage is that the core
can now compute this information in one go, consistently, while holding
only one lock.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>