interfaces/builtin: use permanent snippets in bool-file
This patch changes the security of bool-file to align it with the
permanent vs connected snippet idea. The slot is now permanently able to
control GPIOs (to export/unexport them and to change direction).

The plug side remains the same, that is, plugs can access the GPIO value
or LED brightness, as before.

This also has a nice property of not requiring any restarts of the
service that has this slot when a plug is disconnected because the
plug-side security is not affected by connections. This doesn't matter
yet but will quite soon as interface code is being integrated with the
overlord.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>