snappy: simplify framework application wrapper test
This patch makes a simple framework binary wrapper generator test
explicit, that is not based on an earlier test for non-framework binary
wrapper. The blind text replacements done by this test were confusing
and easily broke with refactoring of the code elsewhere so before
landing the refactoring, I'd like to make this test very easy to follow.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>