interfaces: generate real security file content
This patch builds on the existing foundation to generate non-fake
content of all the files related to security.

The internal security helper interface is extended to contain additional
data for any snap (version, origin and list of apps). This is required
to construct correct content and to locate it in the right directories.

The "xmanager" (read as cross-manager) file is a temporary helper before
all of this moves to the state engine and we can query the state
directly from there. It only exists so that we can ask about the
version, origin and list of apps in a given snap.

Tests are somewhat tweaked since now the real contents of the security
files is very long, a set of mocking functions is provided to replace
the real (very long) seccomp and apparmor header with a shorter version
that is easier to analyze and compare.

Finally, the repository method that writes all the security files (which
is also supposed to move to the interface manager) is extended to pass
the new arguments.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>