testutil,interfaces/udev: simplify udev test code
This patch changes MockCmd and the only test code that was using it,
udev so that testing non-constant mocked behavior is easier. Instead of
an elaborate auto-generated shell script the caller can simply splice a
bit of shell at the end of the mocked command. This has the same power
but greater simplicity.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>