overlord/ifacestate: initial work on Connect()
This patch adds initial work on implementing Connect().

So far everything is detached from the rest of the system (in
particular, there's a second interface repository that this Connect()
operates on.  As the code matures and takes shape it will take over and
replace the copy in the daemon package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>