osutil: make EnsureDirState remove managed files on failure
This patch changes the aforementioned to switch to erase mode when an
error during file-writing stage. This affects how the function is
supposed to be used. Instead of managing all the files in a given
directory or governed by a broad namespace (e.g. "*.snap") the function
is supposed to be used to manage files scoped to a specific snap (e.g.
"docker.*.snap").

This way the erase mode actually helps to attain sanity as to what
happens on errors and errors have isolated impact (other apps are snaps
are not harmed by an error in unrelated snaps)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>