interfaces/apparmor: make combineSnippets private
This patch makes combineSnippets private. To make this happen several
changes to tests had to be made. Most of the test changes are related to
correctly installing/removeing snaps with a interfaces within each test.
Tests for combineSnippets were changed to observe the on-disk state
instead of the snippets in-flight.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>