interfaces: add auto-connection information to interfaces
This patch adds a new method to indicate if a plug / slot of a given
interface should be automatically connected.

Currently only the following interfaces are marked for auto-connection:

 - network
 - network-bind

Other interfaces may be added later, once assertions are in place to
ensure that those are allowed for a given snap (e.g. firewall-control).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>