interfaces: change SecurityBackend.Remove take a string
This patch changes the SecurityBackend interface to take a simple string
in the Remove() method. This matches the similar changes made earlier
today to RemoveSnap() and DisconnectSnap().

The rationale is also that it should be possible to remove security of a
snap with just the name alone. This matches all the other code in the
upcoming doRemoveSecurity handler in the interfaces manager.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>