interfaces/apparmor: always load all apparmor profiles when setting up a snap
This patch ensures that each time we setup apparmor for aspecific snap
we load all profiles. This fixes the unlikely case when a profile is
not changed on disk because earlier setup already put it into place but
snapd was interrupted right before the profile is loaded into the kernel.

The task that contained the security setup will be re-tried and all
profiles will be correctly loaded.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>