Use native systemctl enable instead of our own implementation
We had our own implementation of systemd.Enable() in the past
because ubuntu-device-flash was using this code when the user
installed additional snaps. Because u-d-f could run on pre-systemd
systems we did not want to depend on systemctl for enable.

However it looks like our version is buggy. Plus with 16.04
systemd is everywhere and we will also change the way that
u-d-f --install works for 16.