Add workaround for misbehaving store
SHORT LIVED OMG HACK TO WORKAROUND SERVER BREAKAGE

We should get only a single result when using a search with
a quoted package name. We get totally incorrect results from
the store if we do that. As a workaround we do a search on
the package_name without quotes. This will mean 'http' will
return http,http-server,http-client etc. So we need to manually
filter for exact matches.

Note that this will break once the results are bigger than
the servers page size. Because we have not many snaps in the
store this is not a concern right now.