interfaces/apparmor/template.go: adjust /dev/shm to be more usable
Allow snap-specific access to files and directories in /dev/shm. The previous
implementation of /dev/shm/snap/SNAP_NAME/SNAP_REVISION was problematic
because it didn't handle shm_open() at all and was uncomfortable for developers
to use with open(). This allows file access in /dev/shm for shm_open() and
files in subdirectories for open().