snap: parse confinement property.
Snaps need to be able to specify if they require devmode or if they can
be run confined. This commit adds the parsing part of support for such a
capability, parsing a property called "confinement" that accepts one of
two values: "devmode" or "strict." Any other value results in a parsing
error.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>