Make network-consumer fail with exitcode 1 on error
and adjust the tests to check for cli.ExecCommandErr() when failures
from the consumer are expected. The rational is that on i386 the
syscall "socketcall" is used and if the interface is not available
the network-consumer gets killed because of seccomp violations.

The previous code was checking for error output only and got
confused by non-zero exit states.