cmd/snap: have 'snap list' display helper message on stderr
This is just a follow up for the bug fix linked below. If there are not
snaps to show the helper message is printed to stdout. This can cause
some undesired confusion when coupled with a tool like grep where it
will apparently "find" something while nothing was really available.

This patch changes the helper message to be displayed on stderr .

Fixes: https://bugs.launchpad.net/snappy/+bug/1587445
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>