Move apparmor profile for snap-confine to src/
This patch takes the apparmor profile for snap-confine itself out of the
debian/ directory and into the src/ directory.

In addition, the profile is now generated to understand build-time
configuration as it has to match the final location of the installed
snap-confine executable.

Lastly the name of the profile in the tree is snap-confine.apparmor.in
which is changed on install time to the apparmor-style
usr.lib.snapd.snap-confine (the actual name depends on libexecdir).

This patch paves the way for the upcoming removal of debian/ from the
upstream repository and integration of downstream packaging into spread
tests.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>