asserts: fix GPG key generation parameters
The Creation-Date was off by twelve hours, and is easier to get right
using the seconds= form.  We must also avoid creating a subkey because
we only ever import the primary public key packet (snapd intentionally
does not have a full OpenPGP implementation).

An end-to-end test of this will come later when we have a command-line
tool that signs assertions that we can use from integration tests.
Generating keys in unit tests would be much too slow.