Revert "client: automatically use most powerful socket available."
This reverts commit 40f8c6491527c9faf6d09acf88261e38d09d21f1.

This revert is happening because os.Stat() seems to return information
saying we can access the socket, even though we can't. I'm unable to
determine the cause of this. The only way to determine if we can or
can't connect to the socket is to attempt it, but doing so and failing
due to denials hangs the process. This idea doesn't seem promising.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>