Use fstatfs() and NSFS_MAGIC to decide about setns()
This patch fixes a potentially iffy code that used to open the mount
namespace file /run/snapd/ns/$SNAP_NAME.mnt and then blindly call
sents() on the file descriptor, if the open call succeeded.

Instead, we now fstatfs() the file descriptor and check the resulting
f_type field. If it is NSFS_MAGIC then we know that we should use
setns(). If not we fall back to regular initialization.

All errors from setns() now result in snap-confine death. This should be
more reliable against unexpected bugs.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>