Ensure that parent is alive after installing PR_SET_PDEATHSIG
The child process used for mount namespace capture uses
prctl(PR_SET_PDEATHSIG, ...) to get a SIGINT signal when the parent
process dies. The problem is that by the time that code runs the parent
process may have already died and the signal will never arrive.

To avoid this ensure that the parent process is still alive with
kill(parent, 0). The parent is not interrogated with getppid() but
instead is inherited from the original parent across fork as the child
process is re-parented to the init process (or locally installed
subreaper) when that occurs.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>