revert 'allow mmaping pulseaudio buffers'
From the apparmor.d man page: 'm - allow PROT_EXEC with mmap(2) calls' and
using this with 'w' should not be allowed by the policy. Pulseaudio was
confirmed to only use 'PROT_READ | PROT_WRITE' and not PROT_EXEC and only
showed up on a non-Ubuntu kernel that did not have the required AppArmor
patches for snappy applied.