Split sorting of mountinfo entries
Various parts of the system can come up with nondeterministic ordering.
This results in, e.g. cgroups being ordered randomly from one boot to
another.

The goal of the script is to discard the randomness so it resorted to
sorting but the sort criteria contains random names from various bits of
snap-confine setup.

This patch splits the sorting so that we first discard randomness from
mount_point alone (giving us something that we can sort) and then sorts
the list by a few more factors, after which the remainder of the
de-randomization is performed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>