many: use ConfinementOptions instead of ConfinementType
After the recent change to use ConfinementType across the security
backends an online discussion was started. The culmiation of the
discussion was that this is insufficient to convey devmode if we choose
to apply devmode to classic snaps and instead of a single type we should
pass the three flags in a structure.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>