Generalize mount-ns-layout test to support i386
This patch generalizes the mount-ns-layout test to work on i386. This is
achieved with a small wrapper script that probes the architecture and
translates it to the snappy naming scheme and additional data files for
i386.

The only difference between i386 traces and existing amd64 traces is the
presence of lxcfs entry in /var/lib.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>