disable autopkgtests on ppc64el
The reason is that autopkgtest thinks there is a regression.
But the tests never worked on ppc64el and the one successful run
was a broken autopkgtest script that did not run any test.

Asking to remove the DB entry that autopkgtest on ppc64el for snapd
was not successful so this is needed to stop the system thinking
there is a regression.