Skip i18n test when no "snappy.mo" file is available
The snapd deb from the archive does not contain .mo files, those
are stripped out by the langpack buildd stuff and put into the
the various langpacks.

Therefore this test only makes sense when we build snapd from
the local source. When running against an official snapd deb
or against the core we will not see translations.