cmd: don't reexec on RHEL family
This patch changes how snap commands decide to re-execute. The extra
check looks if we _can_ reexecute in a given distribution. Technically
due to how paths and libraries are used we cannot yet reexecute on
RHEL-family of distributions (fedora, centos, rhel). This restriction
can be lifted when the reexec from core work is fully merged and tested.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>