interfaces/builtin: add /boot/uboot/config.txt access to core-support (#2857)
As discussed in #2488 this moves the config.txt editing into core-support.

This allows us to write a config hook for core to set the options
for the pi2 config.txt. We do it in core for now because we tightly
control this snap and any mistakes in editing config.txt may make
the system unbootable.