tests: do not use core for "All snaps up to date" check (#2872)
The refresh test assumes that core is not refreshable, however a new core can be published during the execution of the suite (very unlikely but possible), making the final check in the refresh test fail.

These changes try to prevent this possibility by comparing the installed core version (from snap list) with the version available for the current channel (from snap info), if they are equal then we try the core refresh (which the test expects to return a message stating that there's no update available). It is still possible that the core snap would be published between the check of the available version and the call to snap refresh, but much more improbable.