interfaces/apparmor: compensate for kernel behavior change
During the end of the zesty development cycle the kernel was changed (as
a bugfix to apparmor bug related to locking files) and now there are
denials when accessing libc.

Jamie Strandboge suggested to add a rule to allow this and John Johansen
confirmed that it should be used by default as the relevant kernel
bugfix is valid and must stay as-is.

Fixes: https://bugs.launchpad.net/snapd/+bug/1666897
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>