cmd/libsnap: use char in sc_string_append_char
This seems to simplify the code a little without losing any generlity.
The values must naturally fit 8-bits (whatever the range is 0..255 or
-128..127) and the compiler can warn about overflows. The biggest
advantage is that it is less tricky to take elements off a "char *" and
shove them to the sc_string_append_char.

Thanks to chipaca for arguing for this change!

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>