many: only tweak core config if hook exists
With the woes that the core snap went through last week (mostly related
to the presence of the configure hook) we are now at a stage where the
configure hook may or may not be there. When the configure hook is not
available in the core snap published on the edge channel then no pull
request can pass the CI system. To unbreak the world we need to make
changes to the code effectively so this patch makes all uses of the
"snap set core" conditional on the configure hook being present.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>