interfaces/mount: compute mount changes required to transition mount profiles (#3050)
This branch adds a function that can be used to look at two mount profiles (as created by the mount backend) and compute a sequence of changes needed to transform a system from one mount profile to another. This will be used as the key part of snap-update-ns to perform modifications to existing mount namespaces.