.travis.yml: remove travis matrix and do a single sequential run (#3085)
With these changes each build will only require one travis slot. Also, if the static tests fail we won't allocate linode resources. The main goal is to only use one travis slot on each PR instead of 3 as we do now, this hopefully will reduce the initial wait time before the jobs start to execute.

During the discussion it was proposed a sequential execution so that we don't allocate linode resources if static or unit test fails, but this would impose an upfront delay in all the PR builds because of the unit test duration (currently about 6min, but this will increase over time). With the approach in this changeset we use a single travis slot and the unit tests are executed as a spread task in parallel with others.