overlord: fix TestEnsureLoopPrune not to be so racy (#3140)
We are seeing many failures that involve the TestEnsureLoopPrune test.
This test is inherently racy as it uses a actual time to wait for
something to happen. In case the machine is loaded heavily and other
processes contend for resources it may not get scheduled in time for
this to happen

This branch contains (now) a more sophisticated approach suggested
and demonstrated by pedronis.