tests: copy snap-confine apparmor profile into testbed
Ever since we started executing snap-confine from the core snap we rely
on a copy of the apparmor profile from the Debian package to generate
the apparmor profile for snap-confine in the core snap.

My recent changes to the snap-confine required an update to the profile
and I noticed that all the test have failed because apparmor profile
in the core snap was not kept up-to-date.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>