cmd/snap-update-ns: add actual implementation
This patch adds a non-dummy implementation of snap-update-ns. There are
still two pieces missing. The function that determines if a mount change
is needed is dummy and always returns true. The mount changes are not
really performed yet as the Perform function is just a stub. The stubs
will be addressed with separate PRs.

All that the tool now does is to print, to stdout, what should be done
instead of actually doing it. Stderr is a bit more noisy, but
essentially explains the same thing with more detail.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>