tests: copy .real profile as .real
The snap-confine apparmor profile comes in two flavours. One is suffixed
contained the .real file it would take precedence and such modifications
copy the .real file without the suffix. If the base snap already
test execution environment in spread we need to copy one or the other so
that snapd sees the updated apparmor profile. There was a bug where we'd
with .real to avoid dpkg bugs, the other is not. When setting up the
would not be used in practice. This patch corrects this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>