Automatically derive the side info in snap prepare-image for --extra-snaps
When `snap prepare-image` is used with --extra snaps the code will
now automatically contact the store and try to derive the side info
for the extra snaps automatically.

This makes it possible to do:
```
$ snap download core
$ snap prepare-image model.assert target-dir --extra-snaps core
```
and still have a non-sideloaded core snap.

The only downside of this approach is that the "contact" information
is missing.