interfaces/browser-support: deny read on squashfs backing files and LVM vg names (#3268)
* interfaces/browser-support: deny r of squashfs backing_file when sandbox: true

Chromium content api when the sandbox is enabled tries to access
virtual/block/loop[0-9]*/loop/backing_file in sysfs, but it operates fine
without them. These files correspond to the squashfs mounts for snaps and
allowing the access constitutes an information leak since the snap would be
able to enumerate the mounted snaps. This is a rather esoteric access and in
the existing policy, only the udisks2 permanent slot policy allows the access
via a glob rule. The udisks2 snap does not plug browser-support and any slot
implementation of udisks should not either (especially not with 'allow-sandbox:
true')

* add one denial more for virtual/block/dm-[0-9]*/dm/name with sandbox: true