interfaces/builtin: distribute code of touching allInterfaces
This patch moves code that registers an interface in allInterfaces, as
well as the code that ensures a given interface is registered into the
modules related to the actual interface. This should lessen the
conflicts around adding a new interface and simplify the task
conceptually.

The only remaining centralized elements are base declaration entries
(per jdstrand's request) and the list of implicit interfaces (which
could be merged into interfaces easily later).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>