interfaces/mount,cmd/snap-update-ns: discard Change.Needed
This patch removes the unimplemented `Change.Needed` method. The method
was designed to inspect the mount namespace, as exposed by the kernel
mountinfo interface, and look for signs that a change has already
occured but was not recorded (e.g. it was constructed by version of
snap-confine earlier than 2.25). In retrospective this feature is very
complex and not really needed as we know exactly what was mounted so we
don't need to guess (using the much more complex kernel interface).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>