tests: Add spread test for timedatectl
Update the time-control-consumer snap and the interfaces-time-control
spread test to ensure that timedatectl can be used to run the set
set-local-rtc subcommand.

We don't actually attempt to set the time in this test because it would
require the timeserver-control interface so that NTP synchronization can
be disabled as `timedatectl set-time` fails when NTP synchronization is
enabled.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>