arch,release: map armv6 correctly
We used GOARCH so far to map out a device we're running on as an ARM
device but that ignores the revision and doesn't detect armv6 devices
properly which we don't have a core snap for at the moment. This lead
to the armhf core snap being pulled on an armv6 device which doesn't
make any sense. This change now uses utsname to get the correct
machine type from the kernel and adjusts the detect architecture if
we're running on an armv6 based device.