tests: dependency packages installed during project prepare
The objective of this change is to provide an environment for all the
tests as similar as possible, so all the dependencies are installed
during the project setup.

The packages are not removed at the end of the project, so it will help
in case the machine needs to be used to debug.

Openvswitch test is moved to manual due to the dependencies install is
stuck trying to install openvswitch-switch package.