Add support for kernel != userland snapd in seccomp
The old C code supported the following cases for compat syscalls:

1. kernel and userspace are both 32 bit (no compat)
2. kernel and userspace are both 64 bit (need to load 32 bit arch for
   32 bit compat)
3. kernel is 64 bit and userspace is 32 bit (need to load 64 bit arch
   for native syscalls)

This commit ensures case (3) is now also supported with the new
go based seccomp code.

Note that (3) in the context of snaps is very strange. When snapd
runs as a 32bit app it will only request 32bit snaps from the
store. So there has to be 32bit snaps that contain 64bit code and
logic that can auto-detect if 64 bit code can be run and that then
makes use of this 64bit code.