store: talk to api.snapcraft.io for assertions
api.snapcraft.io now has a proxy endpoint for the assertions query
needed by snapd, so we can use that to reduce the number of configured
store hostnames.

SNAPPY_FORCE_SAS_URL still exists separately, because I expect it to be
useful for developers working on assertions to be able to point snapd at
a separate snap-assertions-service deployment without having to set up a
snapdevicegw instance as well.