tests: fix for rng-tools service not restarting
As the rng-tools is a forking service, it needs to know which is the
process to watch to restart the service.

This change is adding the information in the service unit configuration
to make sure when /usr/sbin/rngd is not running for any reasonn, then it
is restarted automatically.

This fix should eliminate the lack of entropy issue in 32 bits machines.

ERROR:
https://travis-ci.org/snapcore/snapd/builds/248441911