interfaces: covert framebuffer to commonInterface
It used to be commonInterface but then changed to a custom one so that
udev snippets could be handled. Then those snippets were disabled
because of the bug with udev tagging breaking interfaces not using it
properly. Since this is all being fixed and udev taggning is supported
by commonInterface let's simplify this code.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>