overlord: rely on more conservative ensure interval
On openSUSE tumbleweed running Linux 4.11.8 and golang 1.8.3 the
observed latency between subsequent calls to Ensure in
overlordSuite.EnsureLoopPrune is on the order of ~500ms versus ~50ms on
of the same binary running on top of Ubuntu Artful with Linux 4.11.0 and
golang 1.8.1. This makes the test fail as the sequence of evens it
widely different and things just don't happen in time.

As a workaround, before the root issue is understood better, use a more
conservative prune interval of 200ms (up from 20ms) which makes this test
pass each time.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>