cmd/snap-confine: genearlize apparmor profile for various lib layout
Depending on the distribution at hand we the conceptual "/lib" may be
at /lib or /usr/lib, the "lib" directory may be spelled
lib, lib32, lib64 or even libx32. It may be optionally followed by a
multiarch suffix. This patch applies those rules to all the shared
libraries we need to map and read.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>