cmd/snap-confine,packaging: import snapd-generated policy
This patch allows snap-confie's apparmor profile to be extended with
additional site policy that is generated by snapd. The purpose of the
site policy is so that it can be used to generate some workarounds for
things like $HOME on NFS (the fact of which is not transparent to
confinement, sadly) or encrypted home filesystem. as well as other
quirks that need it specifically in snap-confine's profile (snapd can
already influence profiles of specific snap applications).

It is important to note that the packaging must ensure the directory
/var/lib/snapd/apparmor/snap-confine.d *must* exist or the profile will
fail to compile. I did this for Ubuntu and openSUSE and will coordinate
to make sure this happens in Debian. Other distributions are not
affected as they don't enable AppArmor yet.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>