Use old golang.org/x/crypto/ssh/terminal to build on powerpc again
The recent update of golang.org/x/crypto/ssh/terminal to the latest
revision broke the builds of snapd on powerpc. The reason for this
is that new golang.org/x/crypto/ssh/terminal pulls in x/sys/unix
which does not build on the powerpc platform. But x/sys/unix is
only needed for solaris support of x/crypto/ssh/terminal so we
can ignore that.

This is the easiest way to get working builds on powerpc back.
We could also fork x/crypto/terminal and remove the solaris
support or we could work on adding powerpc support with gccgo
to x/sys/unix (given that powerpc is no longer a supported
platform in ubuntu 17.04+ this is probably not a good investment).