debian: link snap-update-ns statically
This patch passes a few oddly looking flags to the "go build" command to
ensure that the resulting binary, even though it links to libc, is
static. We need to do this to allow snap-update-ns to correctly execute
in an otherwise empty base snap where we cannot rely on the dynamic
linker to do its work.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>