cmd/snap-update-ns: perform SNAP_NAME validation in C
This patch switches the code to fully validate the provided SNAP_NAME in
the C bootstrap code, before go gets a chance to wake up.

In the past the C code just performed very simple checks and relied on
Go to abort and stop execution before we get to the potentially
dangerous mount operations. This was written with the assumption that
snap-update-ns is not a security-critical component because it is
started by snapd and cannot be started by arbitrary user to elevate its
privileges.

With the advent of snap layouts snap-confine will no longer construct
the full mount namespace and will instead perform the early setup that
is harder to from go. After that it will call snap-update-ns which will
process the mount profile and finish the setup process.

With this change snap-update-ns is directly attackable by any user on
the system in case snap-confine is somehow compromised.

To counter this risk, before we run golang code on top of a,
potentially, compromised execution environment, the bootstrap code will
do strict validation of the provided snap name.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>