snap-seccomp: manually resolve socket() call in tests
In libseccomp 2.3.x the "socket" call on i386/s390x returns a
negative syscall number. This breaks our in-kernel tests. This
patch fixes this by manually resolving the syscall to the real
socket syscall (359).

It also errors now when negative syscalls are found, this should
ensure that we always test all our syscalls.