release: add apparmor query and mock APIs
This patch adds new helpers to the release module. They are placed there
out of convenience and proximity to the forced-devmode flag even though
they are not related to os-release in any way.

The remaining apparmor related functions in the release module are now
shims that just use internally cached apparmor data.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>