interfaces/mount: parse x-snapd-mode= mount option
This patch add support function for parsing x-snapd-mode= mount option
out of the fstab-like mount profiles. It is there to support a feature
of snapd mount layouts where snapd can create mount directory
automatically. The directory needs to have a specific mode and
ownership. This patch caters for the mode aspect.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>