cmd/snap-confine: add detection of stale mount namespace
This patch adds *detection* of stale mount namespaces. This is 1/3 of
the required code, as we also need the code that checks if it is safe to
discard a stale namespace as well as code that actually performs the
discard (all somewhat tricky due to apparmor issues).

The detection logic is simple. We know which base snap we would like to
use and we measure it's major:minor device numbers. We then jump into an
existing namespace and measure the major:minor numbers of the device
backing the root filesystem. If they are different then the mount
namespace is stale.

The next branches will bring in detection of stale and *unused*
namespace, thanks to soon-landing freezer cgroup code.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>