cmd/snap: tell translators about arg names and descs req's
in `cmd/snap`'s `lintDesc`, if a description doesn't start with an
uppercase letter, we panic. Similary in its `lintArg`, if an arg name isn't wrapped in <>s, we panic.

This patch adds a comment so that translators are made aware of the
requirement. It also relaxes the panic down to a notice for these
checks (not for the more structural tests done along the same vein)
unless SNAPD_DEBUG or SNAPPY_TESTING is set.

We should really have an integration test that grabs the translations
and checks snap can run against each of them. This is a lot of cases,
so it's not going to be fast (many distros with many lang packs).