cmd/snap-update-ns: add XSnapdEntryID
This patch adds a helper for determining the identifier of a mount
entry. Identifiers are kept in x-snapd.id= mount option. The
identifiers are going to be assigned by snapd and will be based on
stable hashes.

Identifiers can subsequently be used to associate one mount entry with
another (see next patch).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>