partition/ubootenv: don't panic when uboot.env is missing the eof marker
partition/ubootenv's `OpenWithFlags` expects uboot.env's payload to
finish with an eof marker of two zero bytes (which AIUI actually just
means the file is expected to be null-padded to a certain
size). Previously it wouldn't check whether it actually found that
marker before using it to index the payload, resulting in a panic.

This changes things to use the whole payload instead.

An alternative approach would be to error out.