cmd/snap-update-ns: create missing mount points automatically. (#4008)
* cmd/snap-update-ns: expand tests to mock os.{Lstat,MkdirAll,Chown}

The test code can now fake os.Lstat, os.MkdirAll and os.Chown.
This is done in anticipation to changes where a mount change will
create missing parent directory.

* cmd/snap-update-ns: create missing mount points

This patch allows snap-update-ns to create missing mount points automatically.
This allows creating directories in $SNAP_DATA and other writable areas, automatically.

There's no support for ownership and mode other than root.root and 0755 but
that is coming up shortly afterwards.

* cmd/snap-confine: allow snap-update-ns to mkdir under $SNAP_DATA

This apparmor policy change complements the change in snap-update-ns where now
the apparmor profile will allow writes under $SNAP_DATA.

* cmd/snap-update-ns: add SecureMkdirAll

This patch implements a go version of sc_nonfatal_mkpath that very
carefully creates directories and it doesn't traverse any symlinks.

* cmd/snap-update-ns: explain why we are using SecureMkdirAll
* cmd/snap-update-ns: improve tests
* cmd/libsnap: tweak comments
* cmd/snap-update-ns: fix typo
* cmd/snap-update-ns: add freezer support code
* cmd/snap-update-ns: allow snap-update-ns to freeze and thaw
* cmd/snap-update-ns: update apparmor profile for SecureMkdirAt
* cmd/snap-update-ns: freeze snap processes during update

This patch fixes a possible security issue where snap-update-ns is
racing to mount something on "target" and a malicious snap application
process is racing to create a symlink at "target", thus changing the
effective location of the mounted filesystem.

* cmd/snap-update-ns: create bind mount source directory automatically

This patch extends the automatic directory creation feature to create
source directories of bind mounts. This allows two snaps to share a
subset of the data correctly as no hook (which needs a mount namespace
already) can create directory for us.

* tests: add test for content interface mkdir capability

This patch adds a test that newly added content interface features work
as expected.

* cmd/snap-update-ns: reject mounting affecting symlinks

Unfortunately mount cannot be made not to follow symlinks so we take
extra measures to ensure that we don't even use mount with something
that is not a directory. Apart from freezing all the processes we now
lstat the target (for regular mounts) or both source and target (for
bind mounts), to ensure that they are directories and not anything else.

* cmd/snap-update-ns: update stale comment
* cmd/snap-update-ns: add tests for the freezer code
* cmd/snap-update-ns: chown created directory segments

This patch changes SecureMkdirAll to chown each created path segment.
This is necessary as snap-update-ns may run with uid 0 and gid of a
regular user and in result leave behind directories with group ownership
of the calling user.

Because chown is now integrated into SecureMkdirAll the code and tests
were simplified to remove explicit handling of chown mocking.

* cmd/snap-update-ns: fix typo
* interfaces/mount: add OptsToCommonFlags

This patch adds a sibling function to the existing OptsToFlags.
The new function passes unrecognized options back to the caller.

* cmd/snap-update-ns: minimize diff with upcoming overlayfs

This patch pulls in large number of changes (mainly structural code
layout) from the upcoming feature/transparent-overlays branch so that
eventual difference is smaller and easier to review.

The semantic changes are just around testing and a few small details.
Tests are now showing the system call layer more clearly, without mixing
high-level and low-level operations. Utility code is now in a dedicated
file.

* cmd/snap-confine: constrain profile to write directories
* cmd/snap-update-ns: drop TODO comment

The comment is incorrect as all the created directories should actually
be root-owned. Then snap can chown them as necessary.

* cmd/snap-update-ns: document freezeSnapProcesses
* cmd/snap-update-ns: fail if we cannot actually freeze processes

If after waiting the FREEZING state does not turn into FROZEN then abort
the freeze and fail with an informative message.

* cmd/snap-update-ns: thaw partially frozen processes when failing
* cmd/snap-update-ns: increase FREEZING timeout to 3s
* cmd/snap-update-ns: correct wrong comment
* cmd/snap-update-ns: disallow creating relative directories
* cmd/snap-update-ns: use stronger comment about freeze state
* cmd/snap-update-ns: close and ensure all FDs are closed
* cmd/snap-update-ns: use colon to separate error messages
* cmd/snap-update-ns: clarify attack surface
* cmd/snap-update-ns: rewrite comment
* cmd/snap-update-ns: use defer instead of hand-held close
* cmd/snap-update-ns: drop the Impl suffix
* cmd/snap-update-ns: switch to filepath
* cmd/snap-update-ns: clarify why we support only abs paths
* cmd/snap-update-ns: use O_PATH for secureMkdirAt
* cmd/snap-update-ns: use strings.FieldsFunc instead of Split
* Revert "cmd/snap-update-ns: use O_PATH for secureMkdirAt"

We cannot use fchown with a O_PATH-based fd

This reverts commit 07b54a15fa185b2d72605a397aa047a99952f4ad.