cmd/snap-update-ns: fix collection of changes made
This patch fixes a nasty bug that results in wrong list of made changes.
This is caused by how the for / range syntax operates in golang.

Before this patch the list of changes made contains N copies of the
first change. This seems to be caused by golang using a hidden temporary
variable that gets overwritten but has the same address. We add the
address of that variable to the list of changes.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>