interfaces/kmod: treat failure to load module as non-fatal
Since different kernels may not have the requested module, treat any error from
modprobe as non-fatal and attempt any subsequent module loads. Without this,
failure to load a module means failure to connect the interface and the other
security backends.

This issue was found as part of https://github.com/snapcore/snapd/pull/4157
where the broadcom-asic-control and firewall-control interfaces could not be
connected because some of the modules were not available in the test
environments. Importantly, this change is not just for the test environment,
different kernels will have different kernel configs and we need to make sure
that the interface is connectable for the other backends.