cmd/libsnap: fix parsing of empty mountinfo fields
This patch fixes an apparently unexpected kernel behavior whereby
a tmpfs with empty "mount source" (which is customarily set to "none"
by various tools) results in a mountinfo field with two consecutive
spaces. That is, empty fields are just printed as-is.

The parser is adjusted and a new unit test is added to cover the case
where this was discovered.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>