cmd/snap-confine: use helper process, setns at most once
This patch is a follow-up from the discussion between me, jdstrand and
jjohansen. Due to some constraints we must call setns at most once. This
means that we cannot jump in and out of a mount namespace. This patch
adjust ts the logic to have a helper process jump in, look around and
report back.

EDIT: Even with one process the bug looks the same and the same
workaround is kept as a remedy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>