many: add new `snap refresh --amend <snap>` command
We can not currently refresh from a locally installed snapname
to the same name in the store. However users have requested this
feature, e.g. in https://forum.snapcraft.io/t/2542

This PR implements the ability to refresh an unknown local snap
to a store snap of the same name when the `--amend` flag is used.