Saving the snapd-state without compression
Removing the compression option to save execution time.

For the rpi3 the time to save the state with compression is about 141
seconds, and without compression is about 11 seconds. To apply the state
in the reset.sh it is taking about 20 seconds when the state is
compressed and about 10 seconds when it is uncompressed (this operation
is repeated for all the tests).

For a linode machine there is also a big difference. All the numbers are
stored in the following links.

With this change, if we execute all the tests in a single linode server,
the time saves should be about 7 minutes (2.5 seconds aprox by test
executed).

Pi3:
with compression: https://paste.ubuntu.com/26185300/
without compression: https://paste.ubuntu.com/26185446/

Linode ubuntu-16.04-64:
with compression: https://paste.ubuntu.com/26186380/
without compression: https://paste.ubuntu.com/26186378/