overlord/ifacestate: fix disable/enable cycle to setup security
This patch fixes a bug where a disable/enable would leave part of the
security profiles stale. The issue is caused by incorrect bookeeping
of snaps affected by the operation.

InterfaceManager.reloadConnections sets up connections based on the
snap state but it does so without returning any information about
precisely which snaps were affected. In effect the setupProfilesForSnap
function doesn't have the full picture and will ignore any connections
that did not auto-connect at that instant.

To correct this reloadConnections now returns the list of reconnected
snaps and this information is used in conjunction with the rest of
the sources (disconnected snaps, auto-connected snaps).

Fixes: https://bugs.launchpad.net/snapd/+bug/1708703
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>