overlord/snapstate: do a minimal sanity check on containers
This introduces validateContainer, that uses snap.Container's
Walk to do a minimal sanity check of the container in question.

With this, a snap that has an obvious problem due to e.g. a missing or
non-executable app, or an unreadable snap.yaml, will fail to
install. The error message will alert the user to contact the snap
developers, and the system's log will contain additional details meant
for the developers themselves.