interfaces/builtin: add support for software-watchdog interface
Implement software-watchdog interface that allows use of systemd watchdog
functionality. The interfaces opens systemd notification socket for write
access. The path to the socket is specified in NOTIFY_SOCKET environment
variable as described by systemd.exec(5).

Do not hardcode the path to the socket. Since the service will be run by the
same systemd manager as snapd is, infer the path from the NOTIFY_SOCKET in our
environment. Fallback to /run/systemd/notify which is a 'known default'.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>