snap-confine/nvidia: Implement robust approach to NVIDIA files
Instead of our two hard-coded path lists, we'll form relative paths for
all optional elements in the list against common ancestors known for biarch
distributions.

This builds upon the previous work and allows us to replace or grab files
from the subdirectories, and also allows the private NVIDIA libraries to
exist either within the toplevel libdir or within a private directory, as
seen on Fedora.

The flow is changed slightly so that our first primary call for each libdir
is the only mount, and then we follow up with a simple populate call for the
libdir.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>