many: generate and use per-snap snap-update-ns profile
This patch extends the apparmor security backend to create a dedicated
snap-update-ns profile for each snap. That new profile applies to
snap-update-ns when invoked from snap-confine for that specific snap.
The profile exists to express strict mount and write rules that are
required by the layout and avoid the need for generous "you can write
anywhere and mount anywhere" rule.

At this time the mount rules are not any more strict but this will be
easily addressed with subsequent patches that remove generous rules and
add update-ns snippets to compensate.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>