configstate: validate known core.* options
This PR will reject `snap set core config.keys` that are not known to
the running snapds configuration system. This means that unknown
options like in LP:#1746714 will issue clear messages.

If there is a risk with this, i.e. if we want to allow settings in
core that are unknown today (but may become known in the future)
then this PR is clearly not the right one.