interfaces/apparmor: use a helper to set the scope
The spec.AddSnippet function relies on the under-the-hood scope
information that informs it about the set of security tags that is being
operated on. This makes the interface code straightforward and natural
and is a good thing. This patch refactors how that is done and adds the
name of the snap to the set of scope, contextual information.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>