interfaces/apparmor: early support for snap-update-ns snippets
This patch extends the apparmor backend so that interfaces can now use
apparmor snippets that are destined for the snap-update-ns process that
will execute, confined, to initialize or update the mount namespace.

This is designed so that layout code can generate s-u-n snippets that
describe any writable mimics that may need to be constructed so that
security can be tighter.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>